Fix infinite load when using cloudflare
When using Cloudflare, the rate limiter will limit the IP of the Cloudflare proxy server, instead of each individual user. This means legitimate packets get rate limited. If these packets are join room or join server packets, the client will load infinitely.

If another proxy server is used, the header would have to be changed to obtain the correct IP address for the user.